+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import { type Session } from "next-auth";
 import { SessionProvider } from "next-auth/react";
 import { type AppType } from "next/app";
 
-import { api } from "~/utils/api";
+import { api } from "~/lib/api/api";
 
 import Head from "next/head";
 import "~/styles/globals.css";
